Massive update to sensing code on Arduino Boards -- Removed exponential calibration curves, and replaced them with just a plain voltage divider calculation with the actual resistor values for each rail. + Implemented the DFU mode on the Arduino Board + Added a new tool. AtxDfuTool that enables a host machine to comunicate with an Arduino Board on the ATX DFU mode. - Other improvements and code changes.
This is a BETA WIP commit.

+++ AtxWatchdogTools/build_mono_natives.sh
@@ -186,6 +186,18 @@ if [ ! $? -eq 0 ]; then
     exit 1
 fi
 
+echo "atxdfu..."
+if [[ $OSTYPE == darwin* ]]; then
+    libs="${libc_lib}"
+    __gen_image $clr_target "${bin_path}/${build_type}" "atxdfu.exe" "$target_path/atxdfu" libs
+else
+    __gen_image $clr_target "${bin_path}/${build_type}" "atxdfu.exe" "$target_path/atxdfu"
+fi
+if [ ! $? -eq 0 ]; then
+    echo "Error: Failed to generate image."
+    exit 1
+fi
+
 if [ $libc_lib_local_workaround -eq 1 ]; then
     $(rm -f libc.dylib)
     $(rm -f libc.so)
